drop note. move ObservationArray to new file
