+++ rubin_scheduler/scheduler/utils/__init__.py
@@ -1,5 +1,6 @@
 from .comcam_tessellate import *
 from .footprints import *
+from .observation_array import *
 from .sky_area import *
 from .tsp import *
 from .utils import *
